+++ packages/assignment/src/context/themeContext.ts
@@ -1,0 +1,10 @@
+import { createContext, useContext } from "react";
+import { ThemeContextType } from "../types";
+
+export const ThemeContext = createContext<ThemeContextType | null>(null);
+
+export const useTheme = () => {
+	const context = useContext(ThemeContext);
+	if (!context) throw new Error("useTheme must be used within a ThemeProvider");
+	return context;
+};
